DDP processing

+++ floris_files/functions/ddp.py
@@ -5,6 +5,7 @@ import pandas as pd
 import numpy as np
 import os
 import glob
+import copy
 
 from IPython.display import clear_output
 
@@ -14,10 +15,17 @@ class DDP:
     ########### --- GENERAL DATASET TOOLS --- #################
     
     file_extension = ".cnt"
-    frequency = 500 # Hz
-    
+    frequency = 500 # Hz    
     metadata_filenames = ["children.txt", "cdi.txt", "parents.txt", "CODES_overview.txt"]
 
+    # Channels
+    channel_names = ['Fp1', 'AF3', 'F7', 'F3', 'FC1', 'FC5',
+               'T7', 'C3', 'CP1', 'CP5', 'P7', 'P3',
+               'Pz', 'PO3', 'O1', 'Oz', 'O2', 'PO4',
+               'P4', 'P8', 'CP6', 'CP2', 'C4', 'T8',
+               'FC6', 'FC2', 'F4', 'F8', 'AF4', 'Fp2',
+               'Fz', 'Cz']
+    
     channels_ddp_30 = ['O2', 'O1', 'OZ', 'PZ', 'P4', 'CP4', 'P8', 'C4', 'TP8', 'T8', 'P7', 
                    'P3', 'CP3', 'CPZ', 'CZ', 'FC4', 'FT8', 'TP7', 'C3', 'FCZ', 'FZ', 
                    'F4', 'F8', 'T7', 'FT7', 'FC3', 'F3', 'FP2', 'F7', 'FP1']
@@ -29,51 +37,69 @@ class DDP:
                    'P5', 'CP5', 'CP1', 'C1', 'C2', 'FC2', 'FC6', 'C5', 'FC1', 'F2', 'F6', 
                    'FC5', 'F1', 'AF4', 'AF8', 'F5', 'AF7', 'AF3', 'FPZ']
     
-    channel_names = channels_ddp_30
-    
+    channel_names = channel_names    
+
     montage = 'standard_1020'
-    mne_montage = mne.channels.make_standard_montage(montage) 
+    mne_montage = mne.channels.make_standard_montage(montage)
     mne_info = mne.create_info(channel_names, frequency, ch_types='eeg')
     
-    event_dictionary =  {'12', '13', '14', '15', '2', '3', '4', '5', '55', '66', '77', '88'} 
-    
-    incomplete_experiments = []
-    
-    def read_raw(preload = True, verbose=False):
-        
-        #### Concatenate raw files with same id
-        # Show number of files versus number of experiments
-        #raws[0] is modified in-place to achieve the concatenation.
-        #mne.concatenate_raws([experiments_raw[6]])
-        #concat_raw = mne.concatenate_raws([experiments_raw[6], experiments_raw[7]])
+    # Events
+    # Warning, dictionary differs for each participant.
+    event_dictionary_full =  {'12', '13', '14', '15', '2', '3', '4', '5', '55', '66', '77', '88'}
+    event_dictionary_3 = {'standard': 1, 'deviant': 2, 'first_standard': 3}
+    event_dictionary_2 = {'standard': 1, 'deviant': 2}
 
-        return mne.io.read_raw_bdf(raw_path, preload=read_raw, verbose=verbose)
+    standard_list = ['2', '3', '4', '5']
+    deviant_list = ['55', '66', '77', '88']
+    first_standard_list = ['12', '13', '14', '15']
     
-    def get_events_from_raw(self, raw):
-        # events = mne.find_events(raw, verbose=False, min_duration=2/self.frequency)
+    incomplete_experiments = ["8_11", "108_11", "156_11", "164_11", "619_11", "636_11"]
+    
+    @staticmethod
+    def read_raw(raw_paths, preload = True, verbose=False):
+        """
+        Return the raw experiment from multiple .cnt files
+        """
+        raws = []
+        for raw_path in raw_paths:
+            raws.append(mne.io.read_raw_cnt(raw_path, preload=preload, verbose=verbose))
+        return mne.concatenate_raws(raws)
+    
+    def events_from_raw(self, raw):
+        standard_list = ['2', '3', '4', '5']
+        deviant_list = ['55', '66', '77', '88']
+        first_standard_list = ['12', '13', '14', '15']
+
         events, event_dict = mne.events_from_annotations(raw)
-        events_12 = self.group_events_standard(events)
+        events_3 = copy.deepcopy(events)        
 
-        return events, [1]
+        contains_first_standard = False
 
-
-    def group_events_2(events):
+        for key in event_dict:
+            if key in standard_list:
+                events_3[events==event_dict[key]] = 1        
+            if key in deviant_list:
+                    events_3[events==event_dict[key]] = 2
+            if key in first_standard_list:
+                    events_3[events==event_dict[key]] = 3
+                    contains_first_standard = True
+        if contains_first_standard:
+            return events_3, self.event_dictionary_3        
+        else:
+            return events_3, self.event_dictionary_2
+    
+    def group_events_3(self, events, verbose=False):
         """
-        Only returns the standard and deviant events from all events. 
+        Groups all events into standard, first standard, or deviant events.
         """
 
-        events_3 = copy.deepcopy(events)
-        for i in range(len(events)):
-            for newValue, minOld, maxOld in event_conversion_12:
-                condition = np.logical_and(
-                    minOld <= events_12[i], events_12[i] <= maxOld)
-                events_12[i] = np.where(condition, newValue, events_12[i])
-        return events_12
 
-    def create_labels_raw(self, path_label_csv, dataset_directory, ages_directory):
+        return events_3
+
+    def create_labels_raw(self, dataset_directory, ages_directory, path_save_csv = ""):
         """
         This function creates a .csv file with the labels: filename / code / age_group / age_days
-        The labels are saved in 'path_label_csv'.
+        The labels are saved in 'path_save_csv'.
         """
         age_groups = [5, 11, 17, 23, 29, 35, 41, 47]
 
@@ -110,10 +136,11 @@ class DDP:
         age_df = age_df.drop(columns=['age_months', 'age_years']) # age_days is sufficient
         merged_df = pd.merge(cnt_df, age_df, how = 'left', on = ['age_group', 'code'])
         merged_df['age_days'].fillna(merged_df['age_group'] * 30, inplace = True)
-
-        if os.path.exists(path_label_csv):
-            os.remove(path_label_csv)
-        merged_df.to_csv(path_label_csv)
+    
+        if path_save_csv:
+            if os.path.exists(path_label_csv):
+                os.remove(path_label_csv)
+            merged_df.to_csv(path_label_csv)
         
         return merged_df
